+++ src/string/isNotEmpty.js
@@ -2,5 +2,5 @@
 
 export default function isNotEmptyString (val, trimmed = true) {
     if (typeof val !== 'string' && !(val instanceof String)) return false;
-    return (trimmed ? val.trim() : val).length !== 0;
+    return (trimmed === true ? val.trim() : val).length > 0;
 }