+++ string/shorten.js
@@ -5,12 +5,11 @@ Object.defineProperty(exports, "__esModule", {
 });
 exports["default"] = shorten;
 var _is = _interopRequireDefault(require("./is.js"));
-var _isNotEmpty = _interopRequireDefault(require("./isNotEmpty.js"));
 var _isAbove = _interopRequireDefault(require("../number/isAbove.js"));
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { "default": obj }; }
 function shorten(val, length) {
   var postfix = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : '...';
-  if (!(0, _isNotEmpty["default"])(val) || !(0, _is["default"])(postfix) || !(0, _isAbove["default"])(length, 0)) return !1;
+  if (!(0, _is["default"])(val) || !(0, _is["default"])(postfix) || !(0, _isAbove["default"])(length, 0)) return !1;
   if (val.trim().length <= length) return val.trim();
   return "".concat(val.trim().substr(0, length)).concat(postfix);
 }